fix: ajustes pós merge
